Move global attrs to new `nc_attrs` module
Eventually merge the remainder of `set_daily_ncattrs` into this? Or create a new subpackage?
