+++ seaice_ecdr/nc_attrs.py
@@ -1,0 +1,38 @@
+def get_global_attrs() -> dict[str, str]:
+    new_global_attrs = {
+        "title": (
+            "NOAA-NSIDC Climate Data Record of Passive Microwave"
+            "Sea Ice Concentration Version 5"
+        ),
+        "Conventions": "CF-1.10, ACDD-1.3",
+        "program": "NOAA Climate Data Record Program",
+        "cdr_variable": "cdr_seaice_conc",
+        "software_version_id": "tbd",
+        "metadata_link": "tbd",
+        "product_version": "v05r00",
+        "spatial_resolution": "25km",
+        "standard_name_vocabulary": "CF Standard Name Table (v16, 11 October 2010)",
+        "id": "tbd",
+        "naming_authority": "org.doi.dx",
+        "license": "No constraints on data access or use",
+        "summary": (
+            "<tbd> This data set provides a passive microwave sea ice concentration"
+            "climate data record (CDR)...."
+        ),
+        "keywords": "EARTH SCIENCE > CRYOSPHERE > SEA ICE > SEA ICE CONCENTRATION, Continent > North America > Canada > Hudson Bay, Geographic Region > Arctic, Geographic Region > Polar, Geographic Region > Northern Hemisphere, Ocean > Arctic Ocean, Ocean > Arctic Ocean > Barents Sea, Ocean > Arctic Ocean > Beaufort Sea, Ocean > Arctic Ocean > Chukchi Sea, CONTINENT > NORTH AMERICA > CANADA > HUDSON BAY, Ocean > Atlantic Ocean > North Atlantic Ocean > Davis Straight, OCEAN > ATLANTIC OCEAN > NORTH ATLANTIC OCEAN > GULF OF ST LAWRENCE, Ocean > Atlantic Ocean > North Atlantic Ocean > North Sea, Ocean > Atlantic Ocean > North Atlantic Ocean > Norwegian Sea, OCEAN > ATLANTIC OCEAN > NORTH ATLANTIC OCEAN > SVALBARD AND JAN MAYEN, Ocean > Pacific Ocean, Ocean > Pacific Ocean > North Pacific Ocean > Bering Sea, Ocean > Pacific Ocean > North Pacific Ocean > Sea Of Okhotsk",
+        "keywords_vocabulary": "NASA Global Change Master Directory (GCMD) Keywords, Version 7.0.0",
+        "cdm_data_type": "Grid",
+        "project": "NOAA/NSIDC passive microwave sea ice concentration climate data record",
+        "creator_url": "http://nsidc.org/",
+        "creator_email": "nsidc@nsidc.org",
+        "institution": "NSIDC > National Snow and Ice Data Center",
+        "processing_level": "NOAA Level 3",
+        "contributor_name": "tbd",
+        "contributor_role": "tbd",
+        "acknowledgment": "tbd",
+        "source": "tbd",
+        "platform": "tbd",
+        "sensor": "tbd",
+    }
+
+    return new_global_attrs
